requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
base45==0.4.3
bleach==3.3.1
cattrs==1.8.0
cbor2==5.4.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==35.0.0
cwt==1.3.2
decorator==5.1.0
Deprecated==1.2.13
docutils==0.17.1
ecdsa==0.17.0
idna==3.2
importlib-metadata==4.6.3
install==1.3.4
jsonpath-python==1.0.5
jsonpickle==2.0.0
keyring==23.0.1
more-itertools==8.8.0
packaging==21.0
Pillow==8.4.0
pkginfo==1.7.1
pluggy==0.13.1
ply==3.11
py==1.10.0
pyarmor==6.8.1
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.9.9
Pygments==2.9.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==4.4.1
pytest-runner==4.4
python-dateutil==2.8.2
pytz==2021.3
pyzbar==0.1.8
readme-renderer==29.0
requests==2.26.0
requests-cache==0.8.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
tk==0.1.0
toml==0.10.2
tqdm==4.62.0
twine==3.4.2
typing-extensions==3.10.0.2
url-normalize==1.4.3
urllib3==1.26.6
webencodings==0.5.1
wrapt==1.13.2
zipp==3.5.0
